Partially implemented client w/o macros for routing.
This falls apart once you get to different scoping of resources.

It currently works, because the macro is agnostic about opts being
passed into the match. If it gets too much, interpolation doesnt find
anything to do with the param and moves on, if its missing params, it
returns an error with the param (at runtime) that it knows its missing.

In this approach, the route_map has the params built into the map, so
you have to know which params are required ahead of time.

This surfaces in `Client.create/1`. Since we are posting a resource body
the name of the resource is in the payload, so it infers that the name
(which is generally a part of the URL) should be as well. But since this
is a POST its on the collections URL, which doesnt have a name param.

*This needs to refactor routes_test to generate a test for each swagger
operation.*